Changes made to support the implmentation of the task's webapp.
The timeout for the request-reply searches was changed to 5 secs, down from 15 secs.

Missing validation annotations in ProductDTO and UpdateProductDTO were added to throw an exception when price and stock size were null.

A ValidationErrorResponse object was created to correctly handle presentation in the webapp whenever a request failed due to validation.

pom.xml cleanup - commented code removed.